Move reused threshold data and types into seperate files for reuse
