+++ .eslintrc
@@ -7,6 +7,7 @@
     "unicorn/filename-case": "off",
     "unicorn/prefer-node-protocol": "off",
     "unicorn/numeric-separators-style": "off",
-    "no-console": "off"
+    "no-console": "off",
+    "node/no-missing-import": "off"
   }
 }
